Creating entity Company + adding relationship between Company and SalesRep entity.
